labeler/requirements.txt
pillow==5.1.0